Move new fetch to helper file
