+++ src/helpers/fetchHousesData/fetchHousesData.js
@@ -1,3 +1,5 @@
+import fetchHouseMembers from '../fetchHouseMembers/fetchHouseMembers';
+
 const fetchHouseData = async () => {
   try {
     const fetchedData = await fetch('http://localhost:3001/api/v1/houses');
@@ -10,28 +12,5 @@ const fetchHouseData = async () => {
     return new Error('error fetching houses data');
   }
 };
-
-const fetchHouseMembers = arrayOfHouses => {
-  const newHouses = arrayOfHouses.map(async house => {
-    const unresolvedPromises = house.swornMembers.map(async member => {
-      const newMember = await fetch('http://localhost:3001/api/v1/character', {
-        method: 'POST',
-        body: JSON.stringify({ url: member }),
-        headers: {
-          'Content-Type': 'application/json'
-        }
-      });
-      const jsonMember = await newMember.json();
-
-      return jsonMember.name;
-    })
-
-    const swornMembers = await Promise.all(unresolvedPromises);
-
-    return {...house, swornMembers};
-  })
-
-  return Promise.all(newHouses)
-}
 
 export default fetchHouseData;
